Update admin pour correspondre au convention de OSIS
