+++ formation_metier/models/employe_uclouvain.py
@@ -1,3 +1,4 @@
+from django.contrib import admin
 from django.contrib.auth.models import User
 from django.db import models
 from django.db.models import UniqueConstraint, CheckConstraint, Q, TextChoices
@@ -12,14 +13,23 @@ class RoleFormationFareEnum(TextChoices):
 
 class EmployeUCLouvain(models.Model):
     name = models.CharField(max_length=50, blank=False)
-    numberFGS = models.CharField(max_length=8, blank=False)
+    number_fgs = models.CharField(max_length=8, blank=False)
     role_formation_metier = models.CharField(choices=RoleFormationFareEnum.choices, max_length=50,
                                              default=RoleFormationFareEnum.PARTICIPANT, )
     user = models.ForeignKey(User, on_delete=models.CASCADE, null=False)
 
     class Meta:
-        constraints = [UniqueConstraint(fields=['numberFGS'], name='unique_person')]
+        constraints = [UniqueConstraint(fields=['number_fgs'], name='unique_person')]
         permissions = (('access_to_formation_fare', 'Global access to module formation FARE'),)
 
     def __str__(self):
         return f"{self.name}"
+
+
+class EmployeUCLouvainAdmin(admin.ModelAdmin):
+    fieldsets = [('name', {'fields': ['name']}),
+                 ('number_fgs', {'fields': ['number_fgs']}),
+                 ('role_formation_metier', {'fields': ['role_formation_metier']}),
+                 ('user', {'fields': ['user']}),
+                 ]
+    list_display = ('name', 'number_fgs', 'role_formation_metier', 'user')
